+++ test/dummy/app/controllers/gallery/foundations_controller.rb
@@ -1,0 +1,7 @@
+module Gallery
+  class FoundationsController < ApplicationController
+    def show
+      render_catalog_entry(:foundation)
+    end
+  end
+end